Add T3 Code terms and security policies
