+++ apps/mobile/src/features/settings/SettingsRouteScreen.tsx
@@ -537,7 +537,13 @@ function AppSettingsSection() {
   return (
     <SettingsSection title="App">
       <SettingsRow icon="internaldrive" label="Client Storage" target="SettingsClientStorage" />
+      <SettingsRow icon="doc.text" label="Terms of Service" target="SettingsTermsOfService" />
       <SettingsRow icon="doc.text" label="Privacy Policy" target="SettingsPrivacyPolicy" />
+      <SettingsRow
+        icon="checkmark.circle"
+        label="Security Policy"
+        target="SettingsSecurityPolicy"
+      />
       <View className="flex-row items-center gap-4 p-4">
         <SymbolView
           name="info.circle"
