+++ apps/mobile/src/features/settings/components/settings-sheet-targets.ts
@@ -3,4 +3,6 @@ export type SettingsSheetTarget =
   | "SettingsArchive"
   | "SettingsAppearance"
   | "SettingsClientStorage"
-  | "SettingsPrivacyPolicy";
+  | "SettingsPrivacyPolicy"
+  | "SettingsSecurityPolicy"
+  | "SettingsTermsOfService";
